refactor: SubmissionTypeField を共有部品へ抽出（C1作成フォームと共有準備）
VariationCreateForm 内のローカル関数だった提出区分フィールドを _shared へ切り出し、
C1 新規作成フォーム（#351）でも select モードで再利用できるようにする（挙動不変）。
提出区分はバリ単位の不変属性のため入力は新規時のみ（ADR-0045）。
